Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    double x;
-    scanf("%lf",&x);
-    printf("You entered: %lf",x);
+    float x;
+    scanf("%f",&x);
+    printf("You entered: %f",x);
     return 0;
 }
